feat: format code
