+++ components/amenities/delete-dialog.tsx
