+++ components/fuel-types/delete-dialog.tsx
