+++ components/fuel-types/update-dialog.tsx
