+++ components/manufacturers/delete-dialog.tsx
