+++ components/manufacturers/dialog.tsx
