+++ components/manufacturers/update-dialog.tsx
