+++ components/transmissions/delete-dialog.tsx
