+++ components/transmissions/dialog.tsx
